Fix for OPENSTACK-575
When a demo project tried to luanch a vm in a shared subnet created by
admin, the vm went is error state. The reson being, this subnet was not
created by demo and there was no info of the demo user is the VSD. Now,
when he tried to launch a vm, we errored out. The fix for this is, we
pass the tenantID of the tenant that created the subnet in the first place
to the create_vm calls. The demo user can only launch VMs on the subnets
he creates by hemself or subnets created by admin under a shared network.
The fix of passing the tenantID of the tenant that created the subnet
works in both of the above cases. Note, when the demo user launches a vm
in a subnet created by him, we already have created the user_grp for the demo
user at time he creates the subnet. Hence, passing the tenantID of the subnet
created by the demo user in case of he launching a vm in his created subnet
works fine as VSD already has the user_grp info for the demo project.

Change-Id: I23df9d4ca1077685f2a1c3634aa06afa8c1f0578
Closes-Bug: OPENSTACK-575