Fixes bug OPENSTACK-639
user_name became no longer available as parameter in the log formatting. This caused the log to default to the neutron log and not the specified fip_rate_change_log configuration.
Fix is a rewrite of the log initialization to use the neutron ContextAdapter. It should make user_name available.

Change-Id: Ie27587d6de09312baf6174b3271adcd54c804841
Closes-Bug: OPENSTACK-639