Fix for bug openstack 623
When a port was attached to a subnet and it was deleted via
neutron, neutron rejected it but it was getting deleted from vsd.
Added the fix for this.
Also, removed redundant code from delete_subnet method.
The flow is delete the neutron subnet and gateway port in a
transaction.
Then, if the network is external, delete the shared resource and
return.
If it's not an external n/w, go ahead and delete the vsd subnet.
If we reach this part of execution, we can be sure that the
neutron subnet and the g/w port both have been deleted.

Change-Id: I138710d11fcfcad17fadb620e739145772caeda8
Closes-Bug: Openstack-623