Fix for openstack-643
Check if the user is providing a valid application domain for
appd-delete cli. If the user provided the ID of the l3dom in the
vsd, it is just a domain not an applicaiton domain.
TO fix this, now I check for the deployment plcy. It shld be 'zone'
for valid application domains.

Change-Id: I27bf328b027d38250a0a9b91cfb7b1b10dca8cbc
Closes-Bug: Openstack-643