Fix for openstack 651
When a n/w with a l2dom is updated to an external n/w, the port
that was created with the IP of the dhcp server during the l2dom
create was not deleted when this l2dom was removed and a fip pool
subn was created. This fix deletes that port when the n/w is
updated to an external n/w.

Change-Id: I26aabe6f4c572f28c4d944a24502105a4c7ce76e
Closes-Bug: Openstack-651