Fix for OPENSTACK-652 kilo: port-update with security-groups deletes the vm-interface on VSD
This is a by-product of Fix for OPENSTACK-582.

Based on the step mentioned in the bug there are back to back 2 calls to
update_port. First call to update_port is from nova when it boots a VM.
Second is from CLI when only security-group is getting attached to this port.
In that request, device_owner and device_id field are not present.
We check for device_owner set to none but never check for device_id being reset.
Added that check here with this fix.

Change-Id: I920756d1f94eae7d0c66caf0f218a5821210b393