Fixes bug os-655
When an adv subn. was attached to a rtr and port created on it, the
booting of vm on that port was failing. The fix is like for adv. subn.
when they are created, the subn. ID is added in the nuage_l2dom_templ_id
filed while adding the mapping irrespective whether it is l2 or l3.
Use this ID and modified the get_nuage_vport_by_id method. Basically
check for l2 case, it it fails check for l3 case. Tested code for
regular and adv subnets vm creation.

Change-Id: I6cf9351c09c14a74eeafe2968db22c6a49676780
Closes-Bug: OPENSTACK-655