Fixes bug OPENSTACK-668
Default fip rate was set too soon on the fip object.
Therefor it was not possible to see the difference between a request with
rate limiting specifically entered or a request without rate limiting.
When it's not specifically passed in the request, no complaint should be
made when the rate limiting can't be applied.

Change-Id: Ie0b0928a23a62086fd01500b945814a786caf5d8
Closes-Bug: OPENSTACK-668