Fix for OS-670
Added validation for service creation. The service creation was going
through w/o any error when the tcp service was created with src-port
or dest-port as Any. This was creating a service w/o any issue.
However, when this service was used to be associate with a flow, an
error was thrown saying that service shold be a number. prt-range (1-65535)
or '*'. IMO, this error should have been thorwn at the time of service
creation itself. Iy doesn't. So, I have added these checks in the
neutorn cli itself.

Change-Id: I7bc403b2e8b019d7b6ccabdc4b616b2d5ed005be
Closes-Bug: Openstack-670