Fix for openstack-664, openstack-618:
openstack-664: vlan creation should not be allowed for gateways
that belong to a non-default net-partition. Added a check to verify
that the gateway has a default net-partition permission before
creating a vlan.

openstack-618: vlan deletion should not be allowed for gateways
that belong to a non-default net-partition. Added a check to verify
that the gateway has a default net-partition permission before
deleting the vlan.

Change-Id: I452bb7214f679ae0fc891801ca755080266239ab
Closes-Bug: openstack-664