Fixes Openstack-715
_delete_nuage_vport is called for both delete_port and update_port.
update_port in case of a nova delete which causes the device_owner to change.
The problem is in case of deleting the vminterface, the vport is not deleted because
we assume update_port happened from a nova delete.
It is also possible to delete a port which is in use by a VM so in that case we must
also delete the vport and not only the vminterface.

Change-Id: I4d3a14f307e9407aa0101ba1a3bba060e5c9db97
Fixes-Bug: Openstack-715