Fixing OPENSTACK-578, OPENSTACK-580 + adding gitignore
While running rally jobs I discovered that plugin code required
couple of optimization.

1) At many places we were using API get_net_partition_by_name,
especially when we were dealing with default net-partition.
Name is not index in DB and thus query was taking more time.
Keeping track of default net-partition-id is easy and thats what
I have done to change the query to search based on id (primary-key).

2) add_router_interface and remove_router_interface API were doing
lot of ReST operation within the DB subtransaction block which I
could easily moved out of it. This will enable the lock time to be
reduced significantly and it definitely solved the traceback that
we were getting which are reported in 578 and 580.

Change-Id: Id04402429517991e6db721f160cf3c1ad8c1b7a9