Fix for tempest test case test_create_list_show_update_delete_floating_ip
The test case is same in juno and kilo. In kilo is was not caught yet
since there was no validation in update_nuage_vm_vport method. The
validation was there in juno. But in kilo, we create a vport for
port create which was not the case in juno. So, basically asooc.
fips with ports in juno was like a no op for us. However, in kilo
now if we assoc. fip with a vport. and then again assoc. the same fip
with another vport, the validation added in update_nuage_vm_vport
method will be hit and test case fails.
So, now the fix is like if the fip is assoc with a vport with no
interfaces attached, the assoc. is removed and the assoc with the
new port is created.

Change-Id: I0bb69f8536e065a11c1f6dd251c510c09e212c22
Closes-Bug: OPENSTACK-8888