Fix for OS-771
In case of nova interface-detach where only one interface needs to
be deleted iso both interfaces getting deleted and the VM. The reason
for this was that get_ports(based of device_id filter) was called
after the super(update_port) hence, the no. of ports returned were
not correct. Fixed this by getting the correct # of ports attached to
the vm under consideration. no_of_ports is importtant since in the
back end we call delete_vm or delete_interface based on the # of
ports attached.

Change-Id: Ia8dd0f690b1928a5f246a828ce35f2514f4713c9
Closes-Bug: Openstack-771