Fix for OS-776
After the change in the VSD api for vm interface, this bug was
introduced. Now, even if there is a FIP associated with a vport and
a vm interface is attached to this vport, the FIP addr. is no longer
sent in the vm interface response. This affedted the plugin.
Now, to get that information, using the vportId from the vm interface
response, get the vport info and check if a fip is associated at that
vport.

Change-Id: I8d704026572f0db4fc81256cdb1c8e16c4eee274
Closes-Bug: Openstack-776