Fix for OPENSTACK-950
Allow setting NUAGE_VSD_CMS_ID in localrc. If it is not specified, an
existing CMS will be searched for by name. If that one is not found a
new CMS will be created on the VSD and this ID will be set in
nuage_plugin.ini
If NUAGE_VSD_CMS_ID is set in localrc, it will simply be copied into
tempest.conf and no CMS will be looked for or created.

Change-Id: I70315927805d9a991f123d974d2fbd74555022a0
Closes-Bug: OPENSTACK-950