Fix for: OPENSTACK-1107
Kilo>Liberty difference: the subnet extension driver in kilo used to get
the subnet from the original request as input. In Liberty this is changed
to be the subnet after OpenStack actions have updated it. This results in
the mech driver thinking the user entered things like gateway_ip and
allocation_pools in the request. And as a result of that, these values were
incorrectly being validated against VSD.

Also missed some methods that were moved to another class in Liberty.

Change-Id: I4d6f2196469c5a5b6d70a75298720a225f235752
Closes-Bug: OPENSTACK-1107