Fix for OS-1102 Porting in stable/liberty
Added code to restrict the number fip to one per subnet in an
external network.

The check was incorrect. Hence we were allowed to have
2 fips per subnet. Also, the error message in
_validate_shared_resource is applicable only when
the resource is subnet.
When the resource is network -- during the case of
update_network, the existing_subnet should be returned
and the validation is done in update_newtwork()

Change-Id: I0329ba3ac6fe71b939c9cc9a7d307c3fb8ef7f0c
Closes-Bug: Openstack-1102