Fix for OS-1110
Attempting to delte vm's across multiple tenants leads to error. This
was happening since during delte_vms we were sending tenant_id associated
with the subnet to which the port was attached iso the tenant_id assoc.
with the port to which the vm was attached. This has to be the case only
when the vm is booted under a subnet which belongs to a shared network.
Since under a subnet which belongs to a shared n/w  all tenants can create
vms. The deletion in this case, we send the tenant_id assoc. with the subn.
So, two cases can arise here -- case 1: subn is created under shared n/w by
admin or case 2: subn is created under shared n/w by demo. In both cases, we
send tenant_d assoc. with subn during create and delete. In delete case, the
demo user can launch vms only on subnets created by him.

Change-Id: Ie7d1895db791354378e0f3ab0d0e92b6e34d526c
Closes-Bug: Openstack-1110