Fix for os-758: FIP status remains ACTIVE throughout
When the fip is created make status as 'DOWN'
when a fip is assoc. with a port, change status to 'ACTIVE'
when a fip is disassoc. from a port, change status to 'DOWN'
when a port is deleted when it has a fip assoc. with it, mark
the status of fip as 'DOWN'.

Change-Id: I5736b0116ea05f4d7c448c0ba6b2324949d77478
Closes-Bug: Openstack-758