Fix for os-758: FIP status remains ACTIVE throughout
When the fip is created make status as 'DOWN'
when a fip is assoc. with a port, change status to 'ACTIVE'
when a fip is disassoc. from a port, change status to 'DOWN'
when a port is deleted when it has a fip assoc. with it,
mark the status of fip as 'DOWN'.

Change-Id: I894f35c22c320d1deb13e37754b251adb81db4da
Closes-Bug: Openstack-758