Fix for OPENSTACK-1224
When add_router_interface happens with a port_id (so the port was created
before and this is a vport on VSD), the port gets updated with device_owner
network:router_interface. Further in the flow we raise an exception
because we don't support router interface add with existing ports on the
domain. To clean up we call remove_router_interface, upstream code will
then delete the port and our code will not delete the vport because the
device_owner is network:router interface.
Fix is to reset device_owner as part of the cleanup.

Change-Id: I3335b62ebe10bc4009e8cb1152c3c9de9d593375
Closes-Bug: OPENSTACK-1224