Fix for OPENSTACK-1254
When the port has no fixed_ips the code returns early. Because no fixed_ip
> no subnet > nothing on VSD that we know of.
As a result of that, security groups and address pairs (for example) are
not processed by our plugin, and are therefor not part of the port-create
response. This is incorrect with upstream behavior. Ports without fixed_ip
should still return with all their properties.

Change-Id: I419bdfd305ff5be5187b3e16e539ece2f3d62346
Closes-Bug: OPENSTACK-1254