Fix for OPENSTACK-1445
The attribute order of allowed address pairs is different in a GET request
than when doing a PUT request. As a result of that, upstream test
update_port_with_multiple_ip_mac_address_pair was failing because our
plugin was using get_port as response of a update_port (to add nuage
attributes to the port).

I split up the code which adds nuage attributes to the port and call it
separately when doing update_port.

Change-Id: I2b8c964643245539d711a7fec5572969cb278dc0
Closes-Bug: OPENSTACK-1445