Add 'NUAGE_VSD_CMS_NAME' to parameters and test on existence of eth0 file.
NUAGE_VSD_CMS_NAME can be passed in the local.conf of devstack. When
setting this variable, this will be used for creating/looking up the CMS
on the VSD.
Since Ubuntu 15.10 and RHEL7.2 at least, interfaces are named differently
so "eth0" can't be relied upon. The code is enhanced to fetch the MAC in a
more safe way, and when failing it will fallback to cms name "OPENSTACK"

Change-Id: Iaca84e5c095f69024c8dee80a82c244387e4af25
Closes-Bug: OPENSTACK-1487