Do not first delete all policygroups from port when updating securitygroups
When doing a port-update and setting securitygroups, the code used to first
delete all policygroups on the vport and then add any required policygroup
again. Even when the update actually adds a security group and does not
remove any. This causes open connections to break.
The new code will not remove policygroups first anymore. But just update
the vport's policygroups once with the new, correct, values.

Change-Id: Idc13a09ed89121f4677aa038e8f6cf9be54803c9
Closes-Bug: OPENSTACK-1590
Depends-On: Id9a52783fa2420335293bcb572e99cd30fbc0898