Bring is_attr_set back for a port's security groups.
In the average use case, when the user does not specify a SG for the port,
the default security group will be filled in, so this is not needed.
However, in case a user sets portsecurity to False. This behaviour does not
happen and the security_group ID list will still be the
'ATTR_NOT_SPECIFIED' object.

Change-Id: I60452cb8db59bd603b372b78b3dd4cfd869e54d5
Closes-Bug: OPENSTACK-1610