Make the retrieve_by_x methods static.
The way these methods are called changed because you're meant to be able
to write your own custom method, taking the 3 parameters restproxy,
resource and data.
In the 2 default methods restproxy was just taken from the 'self' parameter
but that made creating custom methods difficult as it requires you to write
the method inside the RESTProxyServer class.

So the post method now passes 'self' as 1st parameter but the 2 default
available methods retrieve_by_external_id and retrieve_by_name have to
become static to support this change.

Change-Id: I6d026d7edc092ea81053123998436b99fcd3482b