Retrieve required l3domains for a firewall policy instead of clearing.
Currently, when a Firewall gets updated with a new policy, the old policy
would get cleared of all its linked l3domains. This is a bug because
multiple firewalls can be using the policy, so instead of clearing
l3domains, we must fetch the required l3domains for this policy from the
database.

Change-Id: I42ec0924f78df5deacee71a1399a9d47a4e8cfdf
Closes-Bug: OPENSTACK-1725