Retrieve required l3domains for a firewall policy instead of clearing.
Currently, when a Firewall gets updated with a new policy, the old policy
would get cleared of all its linked l3domains. This is a bug because
multiple firewalls can be using the policy, so instead of clearing
l3domains, we must fetch the required l3domains for this policy from the
database.

Change-Id: I8c598fedbd044d9b1f4926c89393e0bb457b360f
Closes-Bug: OPENSTACK-1725