Upstream moved some provider_net and portbinding constants to neutron-lib
Also need to disconnect router-mapping object from the session during
router delete. This is a behaviour change from upstream for
https://blueprints.launchpad.net/neutron/+spec/enginefacade-switch

If it is not disconnected from the session. The object will loose all
values when the super().delete_router has run. Because the router-mapping
table has ON DELETE CASCADE foreign key to the router table.

Change-Id: I617db3c9e2f9f9e65113777ff1c8f14bf9c95557