Delete FIP after router-delete
When disassociation a FIP in Neutron, previous behaviour of the 
core plugin was to not delete the FIP in VSP, only remove it from the
vPort. In OPENSTACK-1584 this behaviour has changed, the core plugin now
deletes the FIP in L3domain as well.

delete_floatingip still behaved as per pre-1584.
Thus trying to delete the FIP still on the L3 domain.
This bugfixes changes that behaviour to handle the case where the L3 
domain has been deleted.

Summary: Fixed a bug where deleting a floating ip after deleting
the router resulted in an error message.

Change-Id: I932280f5305a88c427979391fcae6682330f5adf
Closes-Bug: OPENSTACK-1704