Retry the securitygroup code on port-create/update
Due to concurrency, it's possible to end up in a scenario where thread A
deletes a policygroup, while in thread B they're trying to use this
policygroup.
This happens while thread A is making the decision to remove the
policygroup, there is no vport currently attached to it. But at this time,
thread B has already seen the policygroup on VSD and intends to use it
instead of making a new policygroup.


Change-Id: I3935c574d31b0ed78efb57daedcb8fe59fbf147d
Closes-Bug: OPENSTACK-1768