Only raise exception requiring vxlan network/segments when vsd-managed.
When creating a vsd-managed subnet (== nuagenet and net_partition have a
value), the subnet is clearly to be processed by this mechanism driver and
we require the network to be vxlan.
When it is os-managed, the mechanism driver will just ignore the creation,
deletion and update of this subnet.

Subsequent port actions will first look for a subnet_mapping in the
database. Because the mechanism driver did not run, no mapping will be
found and the code will return for the port actions.

Change-Id: Id8052c09134c5ff89248b56087e458c6b77c64f4
Closes-Bug: OPENSTACK-1790