Only delete gateway port when there is a subnet mapping.
If there is no subnet mapping and the code reaches this point, it must be
a router:external network, which has no gateway port to delete.
Without this change, the driver actually crashes but neutron doesn't crash
at the API level because it's inside a post_commit method.
Neutron does log stacktraces though.

Change-Id: I03cc29030b9a30fc81cba02a07dbeeb481958855
Closes-Bug: OPENSTACK-1804