Change how we handle remote_group_id in security group rules.
With the old behavior an empty policy group was created when a rule has a
remote_group_id of a security group which does not exist on VSD yet.
Later, when a port was made with this 2nd security group, there is code to
check if policygroup is empty and, if empty, fill it with ACLs.

This caused trouble with concurrency, because when port-create 1 has just
created the policygroup and is about to add ACLs to it. port-create 2 might
find the policygroup, see it's empty and add ACLs to it. Resulting in
duplicate ACLs.

This commit changes the way remote_group_id is handled by no longer
creating empty policy groups, but also adding the ACL rules to it.
As a result of this, the code that checks on empty policygroup and creating
ACLs if so is no longer needed, and this stops the concurrency issue.

Change-Id: I1d1f1090eb6dc9a4be843e148216c485799c52a0
Closes-Bug: OPENSTACK-1574