Fix update firewall rule
Upstream broke update firewall rule, causing it to raise exceptions when
sending an update without a protocol, or when sending an update with new
source/destination ip address and not passing an ip_version.

So our plugin now adds ip_version and destination to the request.
Upstream bug: https://bugs.launchpad.net/neutron/+bug/1656739

This is broken since Ocata only. Also fixed in Ocata but only for V2.

Change-Id: I15f4b581caf624966c82c25dc77d5a19d48b1726
Closes-Bug: OPENSTACK-1934