fix create and delete_port vm-check to binding:host
Alignment was needed for create, update, and delete port with relation to
device_owner vs binding:host checks to determine wether a VM exists or not.
Binding:host is chosen as the main factor to decide (as with Ironic).
LBaaSv2 is special and we have to take care of it during update, because
during delete we are no longer able to know if it's LBaaS port or not. And
delete_nuage_vm needs to know that.

Change-Id: I35ff33ad6773aabdfa4311ba8ba43c4898fcc5dc
Closes-Bug: OPENSTACK-2055
(cherry picked from commit bf4e48d85277ece45a8bc40c357bac53ba012d87)