Fix dualstack subnet in context of router-interface attach and detach
For OPENSATCK-1990:
When a IPV6 subnet is attached to a router and an IPV4
subnet exists in the same network which is not attached
to a router, then upon deletion of IPV4 subnet, we are not
deleting the dualstack subnet on VSD.

For OPENSTACK-2004:
We fix the subnet mapping when creating a subnet
directly on l3 domain. Secondly  we will no longer
create a dhcp port when creating a subnet
directly on l3domain.

Closes-Bug: OPENSTACK-1990, OPENSTACK-2004
Depends-On: I2e1f2e5bc330c2852bb6b7c2fd922fa0de3eaa2d
Change-Id: Iff45d058ec63dee07ea33ef80ee98a9e30cb2204