Support security groups with stateless rules
A Neutron security group is extended to allow to be stateless instead
of the default stateful. If a security group is set to be stateless,
all its rules will be stateless. The statefulness of a security group
can only be toggled if the security group is not associated with any
Neutron port. We added this extension of Neutron security groups to
the NuageAPI service plugin.

Two general tables are added: a table containing parameter names, and
a table containing a parameter name and associated parameter value for
some security group ID. A row `STATEFUL` is added to the first table.
Stateless security groups are saved in the second table with parameter
value `0` for parameter name `STATEFUL`. Stateful security groups are
not saved by design.

Change-Id: I637e58fe7cb3593e47dd56f13eface4a6de2e13a
Depends-On: Idf523d81bd09004e462b7aed03e7ad09d5f17ad7
Closes-Bug: PROD-6791
Closes-Bug: OPENSTACK-2086
Closes-Bug: OPENSTACK-2087
Closes-Bug: OPENSTACK-2091