Fix hotplugging port count as of Queens
To determine wether we need to create a vm or only a vminterface
to connect to an existing vm, we query neutron for ports with
the same device_id. Prior to Queens this returned including the newly
created port (hotplug). As of Queens it only contains pre-existing ports,
changing the threshold for vm_interface creation.

Change-Id: Ie3999aaf6ce0582e1cdf2fcffa40fa32299b8965