Add support for SRIOV accelerated OVS
This change make ovs mech driver to bind the direct port
if user requested --binding-profile '{"capabilities": ["switchdev"]}'
in the direct port, if a user don't request this capability the SR-IOV
legacy NIC mode is used.


Part of this patchset, the experimental flag for l2bridges support
with virtio is eliminated also (support is now always ON).

Change-Id: I11ef2f5eb29b46af0f560b4f717ace3ae154aa70
Closes-Bug: PROD-9044
Closes-Bug: OPENSTACK-2365