Fixes for unnessary API calls while updating the subnet
update_domain_subnet and update_subnet methods were doing a GET call
each time when they called and the returned object was used only
to check whether the name parameter is changed or not. Now we
put a boolean value that indicates change in name parameter so
we don't do any GET calls to update subnet anymore.

Change-Id: I5dad4f9f23ba4db52202c53fb8b7352ab77f76e5
Closes-Bug: OPENSTACK-2436
Closes-Bug: OPENSTACK-2566