Change port state mgmt to use provisioning blocks
Previoslly, port transition to active was happening when
mech driver did bound port. This is broken if nuage ml2
mech driver is deployed next to 3rd party driver which
does port binding. This patchset will make sure driver is
using provisioning blocks and will notify neutron that port
can go to active when wiring of the port is done in VSP.

Change-Id: I8856bd4a782c309ca500c7e936a22284255fe1f8
Closes-Bug: OPENSTACK-2531