Change enterprise extID to name@openstack
In current code, we fill in the ID of the enterprise as the external ID
(enterpriseID@openstack), this is sub-optimal as it takes another PUT
call after the enterprise-creating POST call. It invokes VSD-34554 also.

This patchset changes the external ID to the enterprise *name*
@openstack.

As all lookups happen by the name of enterprise, this patchset is not
affecting those, nor does it, strictly, have any upgrade implications.

Closes-Bug: OPENSTACK-2580
Depends-On: I273eb32915bce88bd0d55eb7b5382574aea15710
Change-Id: Ib4c832d4b42ce14ccd0d302e39bec23199841a08