Expose VRS on actual uplink interface
VRS will be configured to expose the public subnet on the data nic.
This allows FIP traffic to flow between different machines in
the 10.40.0.0/24 subnet.
The controller will also act as the gateway of the public subnet.
If there is a CONTROLLER_IP defined, meaning this is a compute devstack,
this will not happen.

Change-Id: Ie430c99490f72ccd30deb766e0d69dfd02190af8