Clean up VMInterface on concurrent VM create / Port delete
Fix inconsistent state of Openstack-VSD, occuring sometimes
when binding and deleting a port at the same time.
This resulted in a hanging vminterface because no port binding
is yet to be found in Openstack DB.
Now we delete vminterface when Vport has one associated,
even when the port is unbound.

Closes-Bug: OPENSTACK-2797
Depends-On: http://gerrit.an.nuagenetworks.net:8080/497
Change-Id: Ibd7b8c4e39b7363fae1352e73384484c94f49a26