Concurrent create of redirect target NoneType error
When creating a redirect target concurrently, we do not
create the Policygroup rules multiple times, however this
duplicate rules prevention results in a NoneType error, as
we do not return the actual PolicyGroup that was concurrently
created.
Added a lookup for the concurrently created PG.
Added a unittest for this scenario.

Change-Id: Id4996cba2cb0d6abda4fdc4f2825df3c2035d1de
Closes-Bug: OPENSTACK-2865